+++ CHANGELOG.md
@@ -718,3 +718,4 @@ This is a fast forward to v1.6.3 of daviscook477's fork with a few additional ch
 * Fix Potion Lab scroll bounds (kiooeht)
 
 #### dev ####
+* Sensory Stone event text for custom characters (kiooeht)